Fixed .clang-tidy format and fixed CLI argument transform

+++ src/Git2/Oid.cc
@@ -36,7 +36,7 @@ Oid::isZero() const {
 
 std::string
 Oid::toString() const {
-  std::string buf(static_cast<int>(GIT_OID_MAX_HEXSIZE), '\0');
+  std::string buf(GIT_OID_MAX_HEXSIZE, '\0');
   git_oid_tostr(buf.data(), buf.size() + 1, raw);
   return buf;
 }
